Refactores shared macros and generic structures to seperate file
